Support for ALSA 0.5.x loopback, ia-mode dump commands, various bug fixes.

+++ libecasound/src/eca-control-position.cpp
@@ -49,9 +49,9 @@ long int ECA_CONTROL_POSITION::position_in_seconds(void) const {
 }
 
 double ECA_CONTROL_POSITION::length_in_seconds_exact(void) const {
-  return(length_rep / srate_rep);
+  return(static_cast<double>(length_rep) / srate_rep);
 }
 
 double ECA_CONTROL_POSITION::position_in_seconds_exact(void) const {
-  return(curpos_rep / srate_rep);
+  return(static_cast<double>(curpos_rep) / srate_rep);
 }
